iOS app icon
- Directory changes
