+++ iosApp/iosApp.xcodeproj/xcuserdata/checha.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -94,8 +94,8 @@
             filePath = "iosApp/ViewController.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "71"
-            endingLineNumber = "71"
+            startingLineNumber = "73"
+            endingLineNumber = "73"
             landmarkName = "tableView(_:cellForRowAt:)"
             landmarkType = "7">
          </BreakpointContent>
