+++ iosApp/iosApp/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -31,11 +31,13 @@
       "size" : "40x40"
     },
     {
+      "filename" : "ic_diary_120.png",
       "idiom" : "iphone",
       "scale" : "2x",
       "size" : "60x60"
     },
     {
+      "filename" : "ic_diary_180.png",
       "idiom" : "iphone",
       "scale" : "3x",
       "size" : "60x60"
