+++ iosApp/iosApp/en.lproj/Main.strings
@@ -1,4 +1,4 @@
 
 /* Class = "UILabel"; text = "No tasks"; ObjectID = "SZM-pS-YXS"; */
 "SZM-pS-YXS.text" = "No tasks";
-"bIg-Ot-lWh.text" = "Tasks list";
+"bIg-Ot-lWh.text" = "Task list";
